Set code, actual and expected properties on errors
This applies parts of the Node AssertionError convention to the
AssertionError thrown by Referee. The exposed `actual` and `expected`
properties allow test runners to renders diffs. The `code` property is
always set to "ERR_ASSERTION" and `operator` is set to the type and name
of the Referee assertion or refutation.

Refutation do not expose `actual` and `expected` values because the
values are usually equal to each other.