Reduce actual for match and matchJson
When asserting plain objects using match or matchJson, set the "actual"
object to an excerpt only containing the properties specified by the
matcher. If the original "actual" object is used, the diff produced by
test runners shows all properties that are not defined in the matcher as
missing. With this change, only properties that are actually compared
are shown.