Remove referee.count and rework verify
BREAKING: This removes `referee.count`, `referee.resetCount()` and
`referee.verify()`. Note that `verify()` was never released, so
replacing it will not cause any mirgation issues for users.

As a replacement for `verify()`, a new `referee.verifier()` function is
introduced which creates a `verify()` function, exposing `count` and
verifying the assertion counts as `referee.verify()` did before.

This approach moves the `count` state into the verifier function that is
managed by users, allowing us to remove the global `referee.count`
state.